Renaming the test_programs folder
This just made it more clear what programs where stored in that folder, they are not necessary for the individual program, just the broken down algorithms. They are primitive and not fully optimised. They are just the first attempt at the algorithm before I transfer them to my main program and make them cleaner and more efficient.